KNOX-2022 - Splitting up KnoxShellTable and do minor cleanup
